:recycle: Refactor: CallLog 도메인 패키지 구조로 리팩토링 #95
