+++ src/main/java/callprotector/spring/domain/abuse/repository/AbuseLogRepository.java
@@ -1,7 +1,7 @@
 package callprotector.spring.domain.abuse.repository;
 
 import callprotector.spring.domain.abuse.entity.AbuseLog;
-import callprotector.spring.domain.CallLog;
+import callprotector.spring.domain.calllog.entity.CallLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
