+++ src/main/java/callprotector/spring/global/handler/TwilioMediaStreamProcessor.java
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import callprotector.spring.global.client.FastClient;
 import callprotector.spring.global.common.enums.CallTrack;
-import callprotector.spring.service.CallLogService.CallLogService;
+import callprotector.spring.domain.calllog.service.CallLogService;
 import callprotector.spring.service.CallSessionService.CallSessionService;
 import callprotector.spring.service.CallSttLogService.CallSttLogService;
 import callprotector.spring.domain.user.service.UserService;
