+++ src/main/java/callprotector/spring/global/handler/TwilioMediaStreamsHandler.java
@@ -1,7 +1,7 @@
 package callprotector.spring.global.handler;
 
 import callprotector.spring.global.client.FastClient;
-import callprotector.spring.service.CallLogService.CallLogService;
+import callprotector.spring.domain.calllog.service.CallLogService;
 import callprotector.spring.service.CallSessionService.CallSessionService;
 import callprotector.spring.service.CallSttLogService.CallSttLogService;
